+++ src/Test/Unit/Server/Commander.ts
@@ -1,0 +1,17 @@
+import { Commander } from '../../../Server/Commander';
+
+describe("Server/Commander", function () {
+    it("test command", function () {
+        let tested: boolean = false;
+        let commander = new Commander({
+            test: {
+                description: 'test',
+                action: args => {
+                    if (args[0] == 'ttt') tested = true;
+                }
+            }
+        });
+        commander.run('test', ['ttt'], false);
+        expect(true).toBe(tested);
+    });
+});
