+++ src/Test/Unit/Server/CronManager.ts
@@ -1,0 +1,8 @@
+import { CronManager } from '../../../Server/CronManager';
+
+describe("Server/CronManager", function () {
+    it("test command", function () {
+        let tested: boolean = false;
+        let commander = new CronManager();
+    });
+});
