+++ src/Test/Unit/Server/Lib/Log.ts
@@ -1,4 +1,3 @@
-//require('tsconfig-paths/register');
 import { getLevelNumber, isWritableLevel } from '../../../../Server/Lib/Log';
 
 describe("Server/Lib/Log", function() {
