+++ src/Test/Unit/Server/Lib/String.ts
@@ -1,4 +1,3 @@
-//require('tsconfig-paths/register');
 import { distinct } from '../../../../Server/Lib/String';
 
 describe("Server/Lib/String", function() {
