+++ src/Test/Unit/Server/Lib/Url.ts
@@ -1,4 +1,3 @@
-//require('tsconfig-paths/register');
 import { processUrl } from '../../../../Server/Lib/Url';
 
 describe("Server/Lib/Url", function() {
